pyproject.toml
[tool.poetry]
name = "kinopoisk-unofficial-api"
version = "0.1.0"
description = "Simple wrapper of kinopoiskapiunofficial.tech"
authors = ["Nawot <Nawot001@gmail.com>"]
packages = [{include = "kinopoisk" }]
readme = "README.md"

homepage = "https://github.com/Nawot/kinopoisk-unofficial-api"
repository = "https://github.com/Nawot/kinopoisk-unofficial-api"

keywords = ["kinopoisk", "api"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.1"


[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.2"
aiounittest = "^1.4.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.urls]
"Bug Tracker" = "https://github.com/Nawot/kinopoisk-unofficial-api/issues"

